+++ cloudfront-dist/deployment/lib/props/WebStackProps.ts
@@ -2,6 +2,7 @@ import * as core from '@aws-cdk/core';
 
 export interface WebStackProps extends core.StackProps {
     domainName?: string;
+    altDomainName?: string;
     zone?: string;
     FrontAppName?: string;
     version?: string;